+++ src/Extension/AbstractExtension.php
@@ -1,0 +1,60 @@
+<?php
+
+namespace PhpZone\PhpZone\Extension;
+
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
+use Symfony\Component\DependencyInjection\Extension\InvalidArgumentException;
+
+abstract class AbstractExtension implements ExtensionInterface
+{
+    /**
+     * Loads a specific configuration.
+     *
+     * @param array $config An array of configuration values
+     * @param ContainerBuilder $container A ContainerBuilder instance
+     *
+     * @throws InvalidArgumentException When provided tag is not defined in this extension
+     *
+     * @api
+     */
+    abstract public function load(array $config, ContainerBuilder $container);
+
+    /**
+     * Returns the namespace to be used for this extension (XML namespace).
+     *
+     * @return string The XML namespace
+     *
+     * @api
+     */
+    public function getNamespace()
+    {
+        return false;
+    }
+
+    /**
+     * Returns the base path for the XSD files.
+     *
+     * @return string The XSD base path
+     *
+     * @api
+     */
+    public function getXsdValidationBasePath()
+    {
+        return false;
+    }
+
+    /**
+     * Returns the recommended alias to use in XML.
+     *
+     * This alias is also the mandatory prefix to use when using YAML.
+     *
+     * @return string The alias
+     *
+     * @api
+     */
+    public function getAlias()
+    {
+        return get_class($this);
+    }
+}